+++ pcep/pcc-mock/src/main/java/org/opendaylight/protocol/pcep/pcc/mock/protocol/PCCServerPeerProposal.java
@@ -9,7 +9,6 @@ package org.opendaylight.protocol.pcep.pcc.mock.protocol;
 
 import static java.util.Objects.requireNonNull;
 
-import java.math.BigInteger;
 import java.net.InetSocketAddress;
 import org.eclipse.jdt.annotation.NonNull;
 import org.opendaylight.protocol.pcep.PCEPPeerProposal;
@@ -17,12 +16,13 @@ import org.opendaylight.yang.gen.v1.urn.opendaylight.params.xml.ns.yang.controll
 import org.opendaylight.yang.gen.v1.urn.opendaylight.params.xml.ns.yang.controller.pcep.sync.optimizations.rev181109.Tlvs3Builder;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.params.xml.ns.yang.controller.pcep.sync.optimizations.rev181109.lsp.db.version.tlv.LspDbVersionBuilder;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.params.xml.ns.yang.pcep.types.rev181109.open.object.open.TlvsBuilder;
+import org.opendaylight.yangtools.yang.common.Uint64;
 
 public class PCCServerPeerProposal implements PCEPPeerProposal {
     private boolean isAfterReconnection;
-    private final BigInteger dbVersion;
+    private final Uint64 dbVersion;
 
-    public PCCServerPeerProposal(final @NonNull BigInteger dbVersion) {
+    public PCCServerPeerProposal(final @NonNull Uint64 dbVersion) {
         this.dbVersion = dbVersion;
     }
